feat: restrict user management to admin role

+++ src/main/java/com/dunnas/chamados_condominio/application/usecases/user/DeleteUser.java
@@ -1,6 +1,8 @@
 package com.dunnas.chamados_condominio.application.usecases.user;
 
 import com.dunnas.chamados_condominio.application.gateways.UserGateway;
+import com.dunnas.chamados_condominio.domain.entity.Role;
+import com.dunnas.chamados_condominio.domain.entity.User;
 
 public class DeleteUser {
     private final UserGateway userGateway;
@@ -9,7 +11,12 @@ public class DeleteUser {
         this.userGateway = userGateway;
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, String loggedUserEmail) {
+        User loggedUser = userGateway.findUserByEmail(loggedUserEmail);
+
+        if (loggedUser.getRole() != Role.ADMIN) {
+            throw new RuntimeException("Only admins can delete users");
+        }
         userGateway.deleteUser(id);
     }
 }
